refactor: put classes into a subpackage processor
